# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WordForge API
  description: |-
    WordForge 后端 API — v1-stable 端点。

    路径基础路径 `/api`（见 servers[0].url）。
    本规格由 `src/openapi.rs` 自动生成，通过 `cargo test --test openapi_export` 导出，CI 以 `git diff --exit-code docs/openapi.yaml` 防止规格漂移（drift）。
  version: 1.3.0-beta.1
servers:
- url: /api
paths:
...
  /telemetry/app-events:
    post:
      tags:
      - telemetry
      summary: 批量上报客户端埋点命名事件（behavior/error/perf）
      description: m073 埋点事件流。批量 ≤50 条/请求；逐条校验部分成功（errors[] 带 index/code），不整批 400。name 须匹配 ^[a-z0-9_]{1,64}$；props 为 ≤16 键标量对象（字符串 ≤128 字符、序列化 ≤2KB）。clientTsMs 钳制 [now-7d, now+5min]（离线补传窗，宽于 learning 事件的 30min 系有意为之）。(deviceId, clientEventId) 跨请求幂等去重。behavior/perf 受 probe_sampling_config 的 app_behavior/app_perf 行采样，error 恒不采样。须带 X-Device-Id / X-Device-Platform / X-App-Version 头，设备须已注册。
      parameters:
      - name: X-Device-Id
        in: header
        required: true
        schema:
          type: string
      - name: X-Device-Platform
        in: header
        required: true
        schema:
          type: string
      - name: X-App-Version
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppEventsRequest'
        required: true
      responses:
        '200':
          description: '`{accepted, duplicates, sampledOut, failed, errors: [{index, clientEventId, code, message}]}`；限流命中软丢弃 `{accepted: 0, throttled: true}`。逐条错误码：APP_EVENT_INVALID_ID / APP_EVENT_DUPLICATE_ID / APP_EVENT_INVALID_NAME / APP_EVENT_INVALID_CATEGORY / APP_EVENT_INVALID_TS / APP_EVENT_INVALID_PROPS'
          content:
            application/json:
              schema:
                allOf:
                - type: object
                  required:
                  - success
                  properties:
                    success:
                      type: boolean
                - type: object
                  required:
                  - data
                  properties:
                    data:
                      type: object
        '400':
          description: 缺头 — MISSING_DEVICE_ID / MISSING_OS / MISSING_APP_VERSION；X-Device-Id 格式非法 — INVALID_DEVICE_ID；批量超限(>50) — APP_EVENTS_TOO_LARGE
        '401':
          description: 令牌缺失或已过期 — AUTH_UNAUTHORIZED
        '403':
          description: 设备未注册 / 设备归属与当前账号不符 — DEVICE_NOT_REGISTERED / DEVICE_OWNERSHIP_MISMATCH
      security:
      - bearerAuth: []
...
components:
  schemas:
    AppEventsRequest:
      type: object
      required:
      - events
      properties:
        events:
          type: array
          items:
            type: object
            required:
            - clientEventId
            - name
            - category
            - clientTsMs
            properties:
              category:
                type: string
                enum:
                - behavior
                - error
                - perf
              clientEventId:
                type: string
              clientTsMs:
                type: integer
              name:
                type: string
              props:
                type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...